Remove redundant TODO
This is redundant because the default MIDI tempo is 120 BPM, so there's
no need to override it for each instrument's track to match the
instrument state default. Similarly, the default JVM MIDI track volume
of 100/127 is consistent with the default that Alda documents at
https://github.com/alda-lang/alda/blob/master/doc/attributes.md#track-volume,
so again there's no need to set it on each instrument's track.